component: Delta angle utilities. Fix ambiguity in symmetric normalization.
we used to have the SymmetricDeg function which
forced the input angle into [-180, 180]. This is
wrong because we can never tell whether the output of 
a calculation will be + or - 180 degrees.
NormalizeDeg produces [0, 360) so in conformity with that
I made SymmetricDeg now return an angle in [-180, 180), i.e.
the invariant -180 <= x < 180 always holds.

R=bsuter
I=4530090