Recovery policy
Decides how often a recovery action can be triggered ("repair time") and how often it can be triggered for the same root-cause alarm condition ("retry-time").
The second timer should avoid frantic rebooting in the presence of persistent faults which do not seem to get repaired by whatever it is we can do (but
might still be worth trying every few hours or so...). Store & Load recovery policy state (timestamp of last recovery) to a log so that we have a memory
across restarts & reboots.

R=grundmann
I=4633060