Remove homing current limit.
The EPOS 24/5 has a continuous current limit of 5A, the 1s limit is 10A.
According to 
http://www.maxonmotor.es/PDF/motor_ctrl/EPOS%2024-5/Application%20Note%20CANopen%20Basic%20Information.pdf


The object index 0x2080 is the "Current Threshold for Homing Mode". I repeat: for homing mode.
So it makes no sense to increase it.
What we want to set is 0x6410 01 (Continuous Output current limit) to 5000mA and
0x6410 02 Output current limit to 10000mA, but probably this is already the default.